move around files responsible for loading and executing user provided files
