+++ packages/client/src/__tests__/executeCodechecksFile.test.ts
@@ -1,4 +1,4 @@
-import { executeCodechecksFile } from "../codechecksFile";
+import { executeCodechecksFile } from "../file-handling/execution";
 import { join } from "path";
 import { CodechecksJson, executeCodechecksJsonString } from "../file-executors/executeJson";
